# docs/ARCH/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SPOrTS API",
    "version": "1.0.0",
    "description": "Contratti congelati rispetto all'implementazione attuale (core endpoints)"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/api/venues/{id}/announcements": {
      "get": {
        "summary": "Public venue announcements",
        "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],
        "responses": { "200": {"description": "OK"}, "400": {"description": "Bad request"} }
      }
    },
...
    "/api/match-announcements/match/{matchId}/venues": {
      "get": { "summary": "Venues for a match", "parameters": [{"in":"path","name":"matchId","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"} } }
    },
...
    "/api/match-announcements/public/{id}": { "get": { "summary": "Public announcement", "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"}, "404": {"description": "Not found"} } } },
    "/api/match-announcements": { "post": { "security": [{"bearerAuth": []}], "summary": "Create announcement", "responses": { "201": {"description": "Created"}, "400": {"description": "Validation"}, "409": {"description": "Conflict"} } } },
    "/api/match-announcements/venue": { "get": { "security": [{"bearerAuth": []}], "summary": "Venue announcements", "responses": { "200": {"description": "OK"} } } },
    "/api/match-announcements/{id}": { "put": { "security": [{"bearerAuth": []}], "summary": "Update announcement", "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"} } }, "delete": { "security": [{"bearerAuth": []}], "summary": "Delete announcement", "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"} } } },
    "/api/match-announcements/{id}/archive": { "patch": { "security": [{"bearerAuth": []}], "summary": "Archive announcement", "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"} } } },
    "/api/match-announcements/track/match-click": { "post": { "summary": "Track match click", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"matchId": {"type": "string"}}}}}}, "responses": { "200": {"description": "OK"}, "400": {"description": "Bad request"} } } },
...
    "/api/offer-templates": { "get": { "security": [{"bearerAuth": []}], "summary": "List offer templates", "responses": { "200": {"description": "OK"} } }, "post": { "security": [{"bearerAuth": []}], "summary": "Create offer template", "responses": { "201": {"description": "Created"} } } },
    "/api/offer-templates/{id}": { "put": { "security": [{"bearerAuth": []}], "summary": "Update template", "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"}, "404": {"description": "Not found"} } }, "delete": { "security": [{"bearerAuth": []}], "summary": "Delete template", "parameters": [{"in":"path","name":"id","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"}, "404": {"description": "Not found"} } } },
...
  }
}